+++ nfEngine/Renderers/nfRendererVk/CommandRecorder.hpp
@@ -47,25 +47,23 @@ public:
     CommandListID Finish() override;
 
     /// Graphics pipeline methods
-    void SetVertexBuffers(int num, const BufferPtr* vertexBuffers, int* strides, int* offsets) override;
+    void SetVertexBuffers(uint32 num, const BufferPtr* vertexBuffers, uint32* strides, uint32* offsets) override;
     void SetIndexBuffer(const BufferPtr& indexBuffer, IndexBufferFormat format) override;
-    void BindResources(size_t slot, const ResourceBindingInstancePtr& bindingSetInstance) override;
-    void BindVolatileCBuffer(size_t slot, const BufferPtr& buffer) override;
+    void BindResources(uint32 slot, const ResourceBindingInstancePtr& bindingSetInstance) override;
+    void BindVolatileCBuffer(uint32 slot, const BufferPtr& buffer) override;
     void SetResourceBindingLayout(const ResourceBindingLayoutPtr& layout) override;
     void SetRenderTarget(const RenderTargetPtr& renderTarget) override;
     void SetPipelineState(const PipelineStatePtr& state) override;
     void SetStencilRef(unsigned char ref) override;
-    void SetViewport(float left, float width, float top, float height,
-                     float minDepth, float maxDepth) override;
-    void SetScissors(int left, int top, int right, int bottom) override;
-    void Clear(int flags, uint32 numTargets, const uint32* slots, const Math::Float4* colors,
-               float depthValue, uint8 stencilValue) override;
-    void Draw(int vertexNum, int instancesNum, int vertexOffset, int instanceOffset) override;
-    void DrawIndexed(int indexNum, int instancesNum, int indexOffset, int vertexOffset, int instanceOffset) override;
+    void SetViewport(float left, float width, float top, float height, float minDepth, float maxDepth) override;
+    void SetScissors(int32 left, int32 top, int32 right, int32 bottom) override;
+    void Clear(uint32 flags, uint32 numTargets, const uint32* slots, const Math::Float4* colors, float depthValue, uint8 stencilValue) override;
+    void Draw(uint32 vertexNum, uint32 instancesNum, uint32 vertexOffset, uint32 instanceOffset) override;
+    void DrawIndexed(uint32 indexNum, uint32 instancesNum, uint32 indexOffset, int32 vertexOffset, uint32 instanceOffset) override;
 
     /// Compute pipeline methods
-    void BindComputeResources(size_t slot, const ResourceBindingInstancePtr& bindingSetInstance) override;
-    void BindComputeVolatileCBuffer(size_t slot, const BufferPtr& buffer) override;
+    void BindComputeResources(uint32 slot, const ResourceBindingInstancePtr& bindingSetInstance) override;
+    void BindComputeVolatileCBuffer(uint32 slot, const BufferPtr& buffer) override;
     void SetComputeResourceBindingLayout(const ResourceBindingLayoutPtr& layout) override;
     void SetComputePipelineState(const ComputePipelineStatePtr& state) override;
     void Dispatch(uint32 x, uint32 y, uint32 z) override;
