+++ nfEngine/Renderers/nfRendererVk/ResourceBinding.hpp
@@ -10,8 +10,6 @@
 #include "Defines.hpp"
 #include "PipelineState.hpp"
 
-#include <vector>
-#include <map>
 
 
 namespace NFE {
@@ -66,9 +64,9 @@ class ResourceBindingInstance : public IResourceBindingInstance
 
 public:
     bool Init(const ResourceBindingSetPtr& bindingSet) override;
-    bool WriteTextureView(size_t slot, const TexturePtr& texture) override;
-    bool WriteCBufferView(size_t slot, const BufferPtr& buffer) override;
-    bool WriteWritableTextureView(size_t slot, const TexturePtr& texture) override;
+    bool WriteTextureView(uint32 slot, const TexturePtr& texture) override;
+    bool WriteCBufferView(uint32 slot, const BufferPtr& buffer) override;
+    bool WriteWritableTextureView(uint32 slot, const TexturePtr& texture) override;
 };
 
 } // namespace Renderer
